+++ src/60createtable.js
@@ -291,7 +291,7 @@ yy.CreateTable.prototype.execute = function (databaseid, params, cb) {
 
 //	}
 //			if(table.pk) {
-	table.insert = function(r,orreplace) {
+	table.insert = function(r,orreplace,replaceonly) {
 		var table = this;
 
 		var toreplace = false; // For INSERT OR REPLACE
@@ -358,7 +358,7 @@ yy.CreateTable.prototype.execute = function (databaseid, params, cb) {
 			table.update(function(t){
 				for(var f in r) t[f] = r[f];
 			},table.data.indexOf(toreplace),params);
-		} else {
+		} else if(!replaceonly) {
 			table.data.push(r);
 
 		// Final change before insert